% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{caret_step}
\alias{caret_step}
\title{Use caret to calculate model metric}
\usage{
caret_step(d, control_method)
}
\arguments{
\item{d}{data frame as given by \code{\link{subsamples}}}

\item{control_method}{argument passed to \code{\link{trainControl}}
as its "method" argument}

\item{...}{arguments passed to \code{\link{trainControl}}}
}
\value{
Named numeric vector of length two with Rsquared and RMSE
}
\description{
Use caret to calculate model metric
}

