% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Compute R2 and RSME for subsamples of data}
\usage{
metrics(ss, method = "none", N = length(ss))
}
\arguments{
\item{ss}{object as output from \code{\link{subsamples}}}

\item{method}{argument passed to \code{\link{caret_step}}}

\item{N}{number of subamples of each size. This alternative
can be used to reuse a subsamlpe object with larger N set by
\code{\link{subsamples}} instead of generating multiple subsamples
(this \code{N} can not be larger than the original \code{N})
set by \code{\link{subsamples}}).}
}
\value{
object of class "metrics", which is a list of two data.frames,
"Rsquared" and "RMSE where each has columns corresponding
to sample sizes and rows corresponding to repeated samples.
}
\description{
Compute R2 and RSME for subsamples of data
}

