% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{model_step}
\alias{model_step}
\title{Create model and calculate model metric}
\usage{
model_step(d, control_method)
}
\arguments{
\item{d}{data frame as given by \code{\link{subsamples}}}

\item{control_method}{argument passed to \code{\link{trainControl}}
as its "method" argument}

\item{...}{arguments passed to \code{\link{trainControl}}}
}
\value{
Named numeric vector of length two with Rsquared and RMSE
}
\description{
Call \code{\link{train}} if \code{control_method} is anything else than
"none", otherwise call \code{\link{lm}} directly and calculate the metrics.
}

