% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.metrics}
\alias{plot.metrics}
\title{Plot method for class "metrics"}
\usage{
\method{plot}{metrics}(x, ...)
}
\arguments{
\item{x}{object of class "metrics" (see \code{\link{metrics}})}

\item{...}{arguments passed to \code{\link{boxplot}}}
}
\value{
The function is called for its side effects.
}
\description{
Boxplots for each samlpe size for Rsquared and RMSE
}

