% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{r2}
\alias{adj.r2}
\alias{r2}
\alias{unadj.r2}
\title{Convert between adjusted and non adjusted R2}
\usage{
adj.r2(r2, n, p)

unadj.r2(adj.r2, n, p)
}
\arguments{
\item{r2}{(nonadjusted) R2 value}

\item{n}{sample size}

\item{p}{number of parameters in linear model}

\item{adj.r2}{adjusted R2 value}
}
\value{
numeric vector of length one
}
\description{
Convert between adjusted and non adjusted R2
}
\seealso{
\url{https://en.wikipedia.org/wiki/Coefficient_of_determination#Adjusted_R2}
}

