% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data with specified R2}
\usage{
sim_data(r2 = 0.2, n = 1e+06, p = 5, adjusted.r.square = FALSE)
}
\arguments{
\item{r2}{desired adjusted R2 as numeric vector of length one}

\item{n}{number of samples (rows)}

\item{p}{number of dependent variables}

\item{adjusted.r.square}{is the given \code{r2} value adjusted?
(it will then be "unadjusted")}
}
\value{
Object of class "sim_data" wich is basically
a data.frame with first column "Y" and preceding
columns "Xi" (i = 1, ..., p) where
\code{summary(lm(Y ~ ., data = d))$r.squared} is approximately
\code{r2} and with some exrta attributes \code{Rsquared} and
\code{RMSE} with coresponding values based on the whole data.frame
}
\description{
Simulate data with specified R2
}

