% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsamples.R
\name{subsamples}
\alias{subsamples}
\title{Take subsamples from data.frame}
\usage{
subsamples(d, n.sample = seq(50, 500, 50), N = 10)
}
\arguments{
\item{d}{object of class \code{\link{sim_data}}}

\item{n.sample}{subsample sizes}

\item{N}{number of subsamples of each size}
}
\value{
list of length \code{N} with lists of
subsampled data frames for each value of \code{n.sample}
}
\description{
Take subsamples from data.frame
}

